requirements.txt
matplotlib==3.8.0
numpy==1.25.2
Pillow==9.4.0
tqdm==4.65.0
wandb==0.15.8